@@ -12,35 +12,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~2.0.0",
-    "@angular/compiler": "~2.0.0",
-    "@angular/core": "~2.0.0",
-    "@angular/forms": "~2.0.0",
-    "@angular/http": "~2.0.0",
-    "@angular/platform-browser": "~2.0.0",
-    "@angular/platform-browser-dynamic": "~2.0.0",
-    "@angular/router": "~3.0.0",
-    "core-js": "^2.4.1",
-    "moment": "^2.15.1",
-    "rxjs": "5.0.0-beta.12",
-    "ts-helpers": "^1.1.1",
-    "zone.js": "^0.6.23"
+    "@angular/common": "~7.2.12",
+    "@angular/compiler": "~7.2.12",
+    "@angular/core": "~7.2.12",
+    "@angular/forms": "~7.2.12",
+    "@angular/platform-browser": "~7.2.12",
+    "@angular/platform-browser-dynamic": "~7.2.12",
+    "@angular/router": "~7.2.12",
+    "core-js": "^2.6.5",
+    "moment": "^2.24.0",
+    "rxjs": "6.4.0",
+    "rxjs-compat": "^6.4.0",
+    "ts-helpers": "^1.1.2",
+    "tslib": "^1.9.0",
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@types/jasmine": "^2.2.30",
-    "@types/node": "^6.0.42",
-    "angular-cli": "1.0.0-beta.17",
-    "codelyzer": "~0.0.26",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "^7.3.8",
+    "@angular/compiler-cli": "^7.2.12",
+    "@types/jasmine": "^2.8.16",
+    "@types/node": "^10.14.4",
+    "codelyzer": "^5.0.0",
     "jasmine-core": "2.4.1",
     "jasmine-spec-reporter": "2.5.0",
     "karma": "1.2.0",
-    "karma-chrome-launcher": "^2.0.0",
+    "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "^1.0.1",
-    "karma-jasmine": "^1.0.2",
-    "karma-remap-istanbul": "^0.2.1",
+    "karma-jasmine": "^1.1.2",
+    "karma-remap-istanbul": "^0.6.0",
+    "mini-css-extract-plugin": "^0.6.0",
     "protractor": "4.0.9",
-    "ts-node": "1.2.1",
-    "tslint": "3.13.0",
-    "typescript": "2.0.2"
+    "ts-node": "^7.0.1",
+    "tslint": "^5.15.0",
+    "typescript": "^3.2.2",
+    "webpack": "^4.29.6",
+    "webpack-dev-server": "^3.3.1",
+    "webpack-subresource-integrity": "^1.3.2"
   }
 }
