@@ -1,7 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
 import { CalendarService } from './calendar/calendar.service'
@@ -17,7 +17,7 @@ import { CalendarDayComponent } from './calendar/calendar-day/calendar-day.compo
   imports: [
     BrowserModule,
     FormsModule,
-    HttpModule
+    HttpClientModule
   ],
   providers: [CalendarService],
   bootstrap: [AppComponent]
